Add files via upload
I added the inclusion of the virtual function in this example. I felt that I shouldn't create a Polymorphism example without adding the Virtual Function Functionality. The Virtual function is designed to be over-written in the derived classes and I feel that it's important to display this when we are dealing with Polymorphism.